Reorganised ali and sugi's code into one folder
